requirements.txt
numpy
opencv-python
pyautogui
Pillow
keyboard
pyobjc-framework-Quartz